fixed pmd issues
