+++ code/vector3d/ViewModel/src/main/java/ru/unn/agile/vector3d/viewmodel/actions/ActionFactory.java
@@ -1,6 +1,7 @@
 package ru.unn.agile.vector3d.viewmodel.actions;
 
 import ru.unn.agile.vector3d.model.Vector3D;
+import ru.unn.agile.vector3d.viewmodel.actions.impl.*;
 
 public final class ActionFactory {
     private ActionFactory() {
