+++ code/vector3d/ViewModel/src/main/java/ru/unn/agile/vector3d/viewmodel/actions/impl/CrossAction.java
@@ -1,17 +1,18 @@
-package ru.unn.agile.vector3d.viewmodel.actions;
+package ru.unn.agile.vector3d.viewmodel.actions.impl;
 
 import ru.unn.agile.vector3d.model.Vector3D;
 import ru.unn.agile.vector3d.viewmodel.ViewModel;
+import ru.unn.agile.vector3d.viewmodel.actions.IAction;
 
-public class AddAction implements IAction {
+public class CrossAction implements IAction {
     @Override
     public void execute(final ViewModel viewModel) {
         Vector3D vector = new Vector3D(viewModel.getVectorX(),
                 viewModel.getVectorY(), viewModel.getVectorZ());
-        Vector3D addVector = new Vector3D(viewModel.getOtherVectorX(),
+        Vector3D crossVector = new Vector3D(viewModel.getOtherVectorX(),
                 viewModel.getOtherVectorY(), viewModel.getOtherVectorZ());
 
-        viewModel.resultProperty().set(vector.add(addVector).toString());
+        viewModel.resultProperty().set(String.valueOf(vector.cross(crossVector)));
         viewModel.statusProperty().set(ViewModel.Status.SUCCESS.toString());
     }
 }
