+++ code/vector3d/ViewModel/src/main/java/ru/unn/agile/vector3d/viewmodel/validators/impl/VectorValidator.java
@@ -1,6 +1,7 @@
-package ru.unn.agile.vector3d.viewmodel.validators;
+package ru.unn.agile.vector3d.viewmodel.validators.impl;
 
 import ru.unn.agile.vector3d.viewmodel.ViewModel;
+import ru.unn.agile.vector3d.viewmodel.validators.IValidator;
 
 public class VectorValidator implements IValidator {
     @Override
